% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Search}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{W}{knapsack size, positive number}

\item{parallel}{FALSE for non-parallel and default value is FALSE, TRUE for parallel}
}
\value{
The maximum knapsack value and which elements (rows in the data.frame)
}
\description{
Knapsack brute force function that takes a data frame x with two variables v and w, and with a choice of parallel. 
Returns the maximum knapsack value and which elements (rows in the data.frame).
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)

}
\references{
https://en.wikipedia.org/wiki/Knapsack%7D
}
