% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data frame}

\item{W}{knapsack size, positive number}
}
\value{
the maximum knapsack value and which elements (rows in the data.frame)
}
\description{
Knapsack greedy knapsack function that takes a data frame x with two variables v and w. 
Returns the maximum knapsack value and which elements (rows in the data.frame).
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)

}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
