% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Brute force approach for the knapsack problem with non-rcpp optimization}
\usage{
knapsack_brute_force(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Brute force approach for the knapsack problem with non-rcpp optimization
}

