% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force_wo.R
\name{knapsack_brute_force_wo}
\alias{knapsack_brute_force_wo}
\title{Brute force approach for the knapsack problem.}
\usage{
knapsack_brute_force_wo(x, W)
}
\description{
Brute force approach for the knapsack problem.
}

