% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_greedy_rcpp.R
\name{knapsack_greedy_rcpp}
\alias{knapsack_greedy_rcpp}
\title{Greedy Heuristic approach for the knapsack problem with rcpp optimization}
\usage{
knapsack_greedy_rcpp(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
Greedy Heuristic approach for the knapsack problem with rcpp optimization
}

