% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Data frame of knapsack}
\format{
## knapsack_objects
A data frame with 2000 rows and 2 columns:
\describe{
  \item{w}{weight to put item into the knapsack}
  \item{v}{value of the item}
}
}
\usage{
knapsack_objects
}
\description{
Data frame x with two variables v and w for using knapsack algorithm functions.
}
\examples{
library(knapsack)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
\keyword{datasets}
