% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Brute Force Search}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data frame}

\item{W}{knapsack size, positive number}
}
\value{
The maximum knapsack value and which elements (rows in the data.frame)
}
\description{
Knapsack dynamic knapsack function that takes a data frame x with two variables v and w. 
Returns the maximum knapsack value and which elements (rows in the data.frame).
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)

}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem
}
